% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{colprods-package}
\alias{colprods}
\alias{colprods-package}
\title{colprods: calculate column products of numeric matrices}
\description{
A lightweight package containing various functions
    for calculating the column products of numeric matrices. Each
    function performs essentially the same task but each uses a different
    implementation under the hood. The main purpose of this package is to
    provide a simple example to remind myself how to incorporate Rcpp and
    RcppArmadillo functions within a package.
}
\section{Functions}{

\link[colprods]{colprods_arma}\cr
\link[colprods]{colprods_rcpp}\cr
\link[colprods]{colprods_rcpp_rowwise}\cr
\link[colprods]{colprods_apply}\cr
\link[colprods]{colprods_sapply}\cr
\link[colprods]{colprods_for_loop}\cr
}

\examples{
\dontrun{

library(colprods)

# simple example -----

x = matrix(1:9, 3)
colprods_for_loop(x)
colprods_sapply(x)
colprods_apply(x)
colprods_rcpp(x)
colprods_rcpp_rowwise(x)
colprods_arma(x)
as.numeric(colprods_arma(x))

# benchmarking -----

library(microbenchmark)

nr = 1000
nc = 100
x = matrix(rnorm(nr * nc), nr, nc)
microbenchmark(
    colprods_for_loop(x),
    colprods_sapply(x),
    colprods_apply(x),
    colprods_rcpp(x),
    colprods_rcpp_rowwise(x),
    colprods_arma(x),
    as.numeric(colprods_arma(x))
)

nr = 100
nc = 1000
x = matrix(rnorm(nr * nc), nr, nc)
microbenchmark(
    colprods_for_loop(x),
    colprods_apply(x),
    colprods_sapply(x),
    colprods_rcpp(x),
    colprods_rcpp_rowwise(x),
    colprods_arma(x),
    as.numeric(colprods_arma(x))
)
}

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dkidney/colprods}
  \item Report bugs at \url{https://github.com/dkidney/colprods/issues}
}

}
\author{
\strong{Maintainer}: Darren Kidney \email{darrenkidney@yahoo.co.uk}

}
\keyword{internal}
