% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/colprods.r
\name{colprods_arma}
\alias{colprods_arma}
\alias{colprods_rcpp}
\alias{colprods_rcpp_colwise}
\alias{colprods_rcpp_rowwise}
\alias{colprods_apply}
\alias{colprods_sapply}
\alias{colprods_for_loop}
\title{Calculate column products of a matrix}
\usage{
colprods_arma(x)

colprods_rcpp(x)

colprods_rcpp_rowwise(x)

colprods_apply(x)

colprods_sapply(x)

colprods_for_loop(x)
}
\arguments{
\item{x}{a numeric matrix}
}
\description{
Various functions to calculate the column products of a matrix, each with
a different implementation.
}
\examples{
\dontrun{

library(colprods)

# simple example -----

x = matrix(1:9, 3)
colprods_for_loop(x)
colprods_sapply(x)
colprods_apply(x)
colprods_rcpp(x)
colprods_rcpp_rowwise(x)
colprods_arma(x)
as.numeric(colprods_arma(x))

# benchmarking -----

library(microbenchmark)

# tall
nr = 1000
nc = 100
x = matrix(rnorm(nr * nc), nr, nc)
microbenchmark(
    colprods_for_loop(x),
    colprods_sapply(x),
    colprods_apply(x),
    colprods_rcpp(x),
    colprods_rcpp_rowwise(x),
    colprods_arma(x),
    as.numeric(colprods_arma(x))
)

# wide
nr = 100
nc = 1000
x = matrix(rnorm(nr * nc), nr, nc)
microbenchmark(
    colprods_for_loop(x),
    colprods_apply(x),
    colprods_sapply(x),
    colprods_rcpp(x),
    colprods_rcpp_rowwise(x),
    colprods_arma(x),
    as.numeric(colprods_arma(x))
)
}

}
